requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
python-jose[cryptography]==3.3.0
httpx==0.24.1
redis==4.6.0
python-dotenv==1.0.0
